update
重写build配置文件
