--- src/App.less
@@ -1,10 +1,0 @@
-.app {
-  padding: 100px;
-  .demo-1{
-    position: relative;
-    width: 800px;
-    height: 800px;
-    background: #e1e1e1;
-    border: 1px solid #eeeeee;
-  }
-}
