--- src/data.json
@@ -1,60 +1,0 @@
-[
-  {
-    "label": "菜单1",
-    "value": "1",
-    "disabled": true
-  },
-  {
-    "label": "菜单2",
-    "value": "2"
-  },
-  {
-    "label": "菜单3",
-    "value": "3"
-  },
-  {
-    "divider": true
-  },
-  {
-    "label": "菜单4",
-    "value": "4",
-    "subMenu": [
-      {
-        "label": "菜单4-1"
-      },
-      {
-        "label": "菜单4-2",
-        "value": "4",
-        "subMenu": [
-          {
-            "label": "菜单4-2-1"
-          },
-          {
-            "label": "菜单4-2-2"
-          },
-          {
-            "label": "菜单4-2-3"
-          }
-        ]
-      },
-      {
-        "label": "菜单4-3"
-      }
-    ]
-  },
-  {
-    "label": "菜单5",
-    "value": "5",
-    "subMenu": [
-      {
-        "label": "菜单5-1"
-      },
-      {
-        "label": "菜单5-2"
-      },
-      {
-        "label": "菜单5-3"
-      }
-    ]
-  }
-]
