--- src/scorpio-menu/Interface.tsx
@@ -1,9 +1,0 @@
-export interface MenuProps {
-  menuData: object[];
-  position?: {
-    x: number,
-    y: number,
-  };
-  show: boolean;
-  onClose: (e: any) => void;
-}
